+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 ﻿# v0.0.18
 
-_2026-03-19_
+_2026-03-18_
 
 ## What's Changed
+
+### ✨ New Features / 新功能
+
+- install clawhub and skillhub by default
 
 ### 🐛 Bug Fixes / 修复
 
@@ -12,6 +16,10 @@ _2026-03-19_
 ### 🔧 Maintenance / 维护
 
 - bump openclawCompat to >=2026.3.13
+---
+**Full Changelog**: [v0.0.17...v0.0.18](https://github.com/ClawDeckX/ClawDeckX/compare/v0.0.17...v0.0.18)
+
+
 
 ---
 
@@ -197,6 +205,7 @@ _2026-03-15_
 
 
 ---
+